start implementing business logic
