+++ src/main/java/org/mifos/creditbureau/data/CBCreditReportData.java
@@ -1,0 +1,5 @@
+package org.mifos.creditbureau.data;
+//Reformat
+public class CBCreditReportData {
+
+}
