+++ src/main/java/org/mifos/creditbureau/data/CBDestination.java
@@ -1,0 +1,4 @@
+package org.mifos.creditbureau.data;
+//Reformat clientData for the destination credit bureau.
+public class CBDestination {
+}
