+++ src/main/java/org/mifos/creditbureau/data/ClientData.java
@@ -1,0 +1,4 @@
+package org.mifos.creditbureau.data;
+//Reformat data from fineract-client api for internal
+public class ClientData {
+}
