+++ src/main/java/org/mifos/creditbureau/service/CreditBureauConfigurationWriteImplServiceImpl.java
@@ -1,4 +1,20 @@
 package org.mifos.creditbureau.service;
 
-public class CreditBureauConfigurationWriteImplServiceImpl {
+import org.mifos.creditbureau.domain.CreditBureau;
+
+public class CreditBureauConfigurationWriteImplServiceImpl implements CreditBureauRegistrationWriteService {
+    @Override
+    public CreditBureau createCreditBureau() {
+        return null;
+    }
+
+    @Override
+    public void updateCreditBureau() {
+
+    }
+
+    @Override
+    public void updateCreditBureauParams() {
+
+    }
 }
